Move Login and Callback to UserComponents directory
